Run in a local VM cluster

+++ ci/functional/test_main_node.sh
@@ -2,11 +2,16 @@
 
 set -eux
 
-DAOS_TEST_SHARED_DIR=$(mktemp -d -p /mnt/share/)
+if ! DAOS_TEST_SHARED_DIR=$(mktemp -d -p /mnt/share/); then
+    echo "Failed to create temp dir in /mnt/share:"
+    ls -ld /mnt/share
+    id
+    exit 1
+fi
 trap 'rm -rf $DAOS_TEST_SHARED_DIR' EXIT
 
 export DAOS_TEST_SHARED_DIR
-export TEST_RPMS=true
-export REMOTE_ACCT=jenkins
+export TEST_RPMS
+export REMOTE_ACCT
 
 /usr/lib/daos/TESTING/ftest/ftest.sh "$TEST_TAG" "$TNODES" "$FTEST_ARG"
